Add files via upload
this Desktop application is developed in java language which helps the customers to place an order in a coffee shop. this application is developed for a coffee shop where a customer can select coffee according to their preferences and let them a placed order. then the admin will process that order and generates a bill accordingly.